requirements.txt
Flask==3.0.0
psutil==5.9.8
python-dotenv==1.0.1
prometheus-client==0.19.0